refactor(cco): segmented allocator + size==0 + recommended granularity; fold Window into cco_init
Two related changes bundled together — both touch CcoMemAlloc/Free and
the surrounding host-side cco translation unit:

1. Replace bump-pointer slot allocator with a segmented one ported from
   NCCL ncclSpace.

   * Tracks allocated [lo, hi) intervals in a sorted `cuts` array;
     segments alternate empty<->full. First-fit on alloc, shrink/split/
     merge on free. Algorithm in src/cco/cco_init.cpp:
       - SpaceInsertSegment: insert two cuts + dedup adjacent equals
         (zero-length collapse, fuses neighbours).
       - SpaceAlloc / SpaceFree: short wrappers around the above with
         the standard NCCL three-case fast paths.

   * Replaces CcoComm::nextOffset with CcoComm::AllocSpace { cuts }.
     Failure-path rollback in CcoMemAlloc becomes a clean SpaceFree
     instead of the previous best-effort bump-pointer hack. CcoMemFree
     now actually returns the slot to the AllocSpace, eliminating the
     hole-leak in long-running alloc/free churn.

   * test_cco_host adds a Phase 1.6 that exercises:
       - size==0 alloc returning nullptr,
       - alloc → free → alloc returning the SAME VA (slot reuse).
     Verified across 8 ranks.

2. CcoMemAlloc: size==0 returns *outPtr=nullptr immediately (matches
   NCCL ncclMemAlloc semantics); CcoMemFree(nullptr) is a no-op.
   Avoids hipMemCreate(0) and the resulting error paths.

3. CcoCommCreate: query granularity with hipMemAllocationGranularity-
   Recommended (typically 2 MiB) instead of Minimum (typically 4 KiB).
   CCO's usage pattern is "few large windows", so the small internal
   fragmentation buys fewer page-table entries / better TLB hit rate.

4. Folded cco_memory.cpp (CcoWindowRegister × 2 + CcoWindowDeregister)
   into cco_init.cpp. The two files shared all major includes and the
   CcoComm host-side state machine; keeping them separate added a hop
   without helping comprehension. cco_init.cpp is now the single
   translation unit for the host API.

CMake: src/cco/CMakeLists.txt MORI_CCO_SOURCES drops cco_memory.cpp.

Tests (8× MI300X, bnxt_re):
  - test_cco_host 8 ranks (with new size==0 + slot-reuse coverage):
    8/8 PASS
  - test_cco_multiprocess fork 8: 8/8 PASS
  - test_cco_multiprocess MPI 8: 8/8 PASS
  - MORI_ENABLE_SDMA=1 pytest test_dispatch_combine_async_ll.py:
    68 passed, 98 skipped (shmem regression guard).
