+++ src/cco/cco_init.cpp
@@ -2,11 +2,14 @@
 // MIT License — see LICENSE for details.
 #include "mori/cco/cco_api.hpp"
 
+#include <algorithm>
 #include <cstring>
+#include <string>
 #include <vector>
 
 #include "hip/hip_runtime_api.h"
 #include "mori/application/bootstrap/local_bootstrap.hpp"
+#include "mori/application/transport/rdma/rdma.hpp"
 #include "mori/application/transport/sdma/anvil.hpp"
 #include "mori/application/utils/check.hpp"
 #include "mori/utils/hip_compat.hpp"
@@ -20,6 +23,85 @@ static constexpr size_t INTERNAL_SYNC_BYTES = INTERNAL_SYNC_COUNT * sizeof(uint6
 
 static size_t AlignUp(size_t x, size_t align) {
   return (x + align - 1) & ~(align - 1);
+}
+
+/* ========================================================================== */
+/*  AllocSpace — segmented interval allocator (ported from NCCL ncclSpace).   */
+/*                                                                            */
+/*  Tracks allocated [lo, hi) intervals in a sorted `cuts` array. Segments    */
+/*  between consecutive cuts alternate empty<->full. The trailing segment is  */
+/*  always empty (the unallocated frontier). Knowing parity:                  */
+/*    isFull(i) = (i % 2 != cuts.size() % 2)                                  */
+/*                                                                            */
+/*  Alloc is first-fit; Free shrinks/splits/merges. Caller holds the mutex.   */
+/* ========================================================================== */
+
+// Insert two cuts [lo, hi) before `index`, then dedup adjacent equal entries.
+// Adjacent equal cuts collapse zero-length segments and fuse the neighbours.
+static void SpaceInsertSegment(CcoComm::AllocSpace& s, int index, int64_t lo, int64_t hi) {
+  s.cuts.insert(s.cuts.begin() + index, {lo, hi});
+
+  int count = static_cast<int>(s.cuts.size());
+  int r = index, w = index;
+  int64_t prev = (r == 0) ? 0 : s.cuts[r - 1];
+  while (r < count) {
+    int64_t cur = s.cuts[r++];
+    s.cuts[w++] = cur;
+    if (prev == cur) {
+      w -= (w == 1) ? 1 : 2;
+      cur = 0;
+    }
+    prev = cur;
+  }
+  s.cuts.resize(w);
+}
+
+// First-fit search over empty segments. On success writes the chosen aligned
+// offset into *outOffset and returns 0; returns -1 if nothing fits.
+static int SpaceAlloc(CcoComm::AllocSpace& s, int64_t limit, int64_t size,
+                      int64_t align, int64_t* outOffset) {
+  int count = static_cast<int>(s.cuts.size());
+  int i = count % 2;  // First empty segment ends at cuts[i].
+  while (i <= count) {
+    int64_t lo = (i == 0)     ? 0     : s.cuts[i - 1];
+    int64_t hi = (i == count) ? limit : s.cuts[i];
+    int64_t off = (lo + align - 1) & ~(align - 1);
+    if (off + size <= hi) {
+      *outOffset = off;
+      if (i == 0 || off + size == hi) {
+        SpaceInsertSegment(s, i, off, off + size);
+      } else {
+        s.cuts[i - 1] = off + size;  // Extend left-adjacent full segment.
+      }
+      return 0;
+    }
+    i += 2;
+  }
+  return -1;
+}
+
+// Free [offset, offset+size). Returns 0 on success, -1 if not a valid full
+// segment slice.
+static int SpaceFree(CcoComm::AllocSpace& s, int64_t offset, int64_t size) {
+  int count = static_cast<int>(s.cuts.size());
+  if (count == 0 || s.cuts[count - 1] <= offset) return -1;
+
+  int i = 1 - (count % 2);  // First full segment ends at cuts[i].
+  while (i < count && s.cuts[i] <= offset) i += 2;
+  if (i >= count) return -1;
+
+  int64_t lo = (i == 0) ? 0 : s.cuts[i - 1];
+  int64_t hi = s.cuts[i];
+  if (offset < lo || hi < offset + size) return -1;
+
+  if (i != 0 && lo == offset && offset + size != hi) {
+    s.cuts[i - 1] = offset + size;
+  } else if (lo != offset && offset + size == hi) {
+    s.cuts[i] = offset;
+  } else {
+    SpaceInsertSegment(s, i, offset, offset + size);
+  }
+  return 0;
 }
 
 /* ========================================================================== */
@@ -134,9 +216,12 @@ int CcoCommCreate(application::BootstrapNetwork* bootNet, size_t perRankVmmSize,
   allocProp.location.type = hipMemLocationTypeDevice;
   allocProp.location.id = currentDev;
 
+  // RECOMMENDED granularity (typically 2 MiB on modern GPUs) trades a small
+  // amount of internal fragmentation for fewer page-table entries, matching
+  // CCO's "few large buffers" usage pattern.
   size_t granularity = 0;
   HIP_RUNTIME_CHECK(
-      hipMemGetAllocationGranularity(&granularity, &allocProp, hipMemAllocationGranularityMinimum));
+      hipMemGetAllocationGranularity(&granularity, &allocProp, hipMemAllocationGranularityRecommended));
   comm->vmmGranularity = granularity;
 
   // Flat VA covers the LSA team only. Cross-node peers don't use VA — RDMA
@@ -239,36 +324,44 @@ int CcoCommDestroy(CcoComm* comm) {
 /* ========================================================================== */
 
 int CcoMemAlloc(CcoComm* comm, size_t size, void** outPtr) {
+  if (outPtr == nullptr) {
+    MORI_SHMEM_ERROR("CcoMemAlloc: outPtr is NULL");
+    return -1;
+  }
+  if (size == 0) {
+    *outPtr = nullptr;
+    return 0;
+  }
+
   int currentDev = 0;
   HIP_RUNTIME_CHECK(hipGetDevice(&currentDev));
 
   size_t alignedSize = AlignUp(size, comm->vmmGranularity);
 
-  // Reserve a slot via bump pointer; rolled back below if any later step fails.
-  size_t slotOffset;
+  // Reserve a slot via first-fit in the per-rank AllocSpace.
+  int64_t slotOffset64 = 0;
   {
     std::lock_guard<std::mutex> lock(comm->allocMutex);
-    slotOffset = comm->nextOffset;
-    if (slotOffset + alignedSize > comm->perRankSize) {
+    if (SpaceAlloc(comm->allocSpace, static_cast<int64_t>(comm->perRankSize),
+                   static_cast<int64_t>(alignedSize),
+                   static_cast<int64_t>(comm->vmmGranularity),
+                   &slotOffset64) != 0) {
       MORI_SHMEM_ERROR(
-          "CcoMemAlloc: slot exhausted (offset {} + size {} > perRankSize {}). "
-          "Increase perRankVmmSize at CcoCommCreate or reduce window count.",
-          slotOffset, alignedSize, comm->perRankSize);
+          "CcoMemAlloc: slot exhausted (no contiguous {} bytes free in perRankSize={}). "
+          "Increase perRankVmmSize at CcoCommCreate or free unused allocations.",
+          alignedSize, comm->perRankSize);
       return -1;
     }
-    comm->nextOffset += alignedSize;
   }
+  size_t slotOffset = static_cast<size_t>(slotOffset64);
 
   MORI_SHMEM_TRACE("CcoMemAlloc: rank={} size={} alignedSize={} slotOffset={}", comm->rank,
                    size, alignedSize, slotOffset);
 
-  // Best-effort slot rollback: only undoes the bump if no other thread has
-  // moved past us. Holes are leaked until comm destruction (no freelist yet).
+  // Return the reserved slot to the AllocSpace on any failure after this point.
   auto rollbackSlot = [&]() {
     std::lock_guard<std::mutex> lock(comm->allocMutex);
-    if (comm->nextOffset == slotOffset + alignedSize) {
-      comm->nextOffset = slotOffset;
-    }
+    (void)SpaceFree(comm->allocSpace, slotOffset64, static_cast<int64_t>(alignedSize));
   };
 
   hipMemAllocationProp allocProp = {};
@@ -345,8 +438,11 @@ int CcoMemAlloc(CcoComm* comm, size_t size, void** outPtr) {
 /* ========================================================================== */
 
 int CcoMemFree(CcoComm* comm, void* ptr) {
-  // Snapshot meta under the lock, then drop it before the (potentially slow)
-  // hipMem* calls so concurrent MemAlloc isn't blocked.
+  if (ptr == nullptr) return 0;
+
+  // Snapshot meta + return the slot to AllocSpace under the lock, then drop
+  // it before the (potentially slow) hipMem* calls so concurrent MemAlloc
+  // isn't blocked.
   CcoComm::AllocMeta meta;
   {
     std::lock_guard<std::mutex> lock(comm->allocMutex);
@@ -357,6 +453,8 @@ int CcoMemFree(CcoComm* comm, void* ptr) {
     }
     meta = it->second;
     comm->allocTable.erase(it);
+    (void)SpaceFree(comm->allocSpace, static_cast<int64_t>(meta.slotOffset),
+                    static_cast<int64_t>(meta.size));
   }
 
   size_t alignedSize = meta.size;
@@ -393,9 +491,339 @@ int CcoMemFree(CcoComm* comm, void* ptr) {
 
   if (meta.shareFd >= 0) close(meta.shareFd);
 
-  // slotOffset is not returned to nextOffset — bump pointer only. Long-running
-  // alloc/free churn will eventually exhaust perRankSize; a segmented
-  // allocator is future work.
+  return 0;
+}
+
+/* ========================================================================== */
+/*                         CcoWindowRegister (ptr)                         */
+/* ========================================================================== */
+
+int CcoWindowRegister(CcoComm* comm, void* ptr, size_t size, CcoWindow_t* outWin) {
+  auto it = comm->allocTable.find(ptr);
+  if (it == comm->allocTable.end()) {
+    MORI_SHMEM_ERROR("CcoWindowRegister: ptr {} not in allocTable", ptr);
+    return -1;
+  }
+
+  auto& meta = it->second;
+  size_t slotOffset = meta.slotOffset;
+  int shareFd = meta.shareFd;
+  void* localPtr = ptr;
+  int worldSize = comm->worldSize;
+  int rank = comm->rank;
+
+  size_t alignedSize = meta.size;
+
+  MORI_SHMEM_TRACE("CcoWindowRegister: rank={} ptr={} size={} slotOffset={}", rank, ptr, size,
+                   slotOffset);
+
+  int currentDev = 0;
+  HIP_RUNTIME_CHECK(hipGetDevice(&currentDev));
+
+  // P2P: exchange dma-buf FDs with same-node peers and map their slots into
+  // the LSA flat VA.
+  std::vector<int> p2pPeers;
+  for (int pe = 0; pe < worldSize; pe++) {
+    if (comm->ctx->CanUseP2P(pe)) {
+      p2pPeers.push_back(pe);
+    }
+  }
+
+  if (!p2pPeers.empty()) {
+    std::vector<int> sortedGroup = p2pPeers;
+    sortedGroup.push_back(rank);
+    std::sort(sortedGroup.begin(), sortedGroup.end());
+
+    int myPeerRank = 0;
+    for (int i = 0; i < static_cast<int>(sortedGroup.size()); i++) {
+      if (sortedGroup[i] == rank) { myPeerRank = i; break; }
+    }
+    int p2pWorldSize = static_cast<int>(sortedGroup.size());
+
+    // Socket path must agree across the group but be unique per (group, window).
+    // groupId = rank 0's pid; slotOffset identifies the window.
+    std::string socketPath = "/tmp/mori_cco_" + std::to_string(comm->groupId) + "_" +
+                             std::to_string(slotOffset) + "_";
+
+    // Best-effort cleanup of stale sockets from crashed runs.
+    if (myPeerRank == 0) {
+      for (int i = 0; i < p2pWorldSize; i++) {
+        for (int j = 0; j < p2pWorldSize; j++) {
+          std::string stale = socketPath + std::to_string(i) + "_" + std::to_string(j);
+          unlink(stale.c_str());
+        }
+        unlink((socketPath + "barrier_arrive_" + std::to_string(i)).c_str());
+        unlink((socketPath + "barrier_depart_" + std::to_string(i)).c_str());
+      }
+    }
+
+    application::LocalBootstrapNetwork localBoot(myPeerRank, p2pWorldSize, socketPath);
+    localBoot.Initialize();
+
+    std::vector<int> myFds = {shareFd};
+    std::vector<std::vector<int>> allFds;
+    if (!localBoot.ExchangeFileDescriptors(myFds, allFds)) {
+      MORI_SHMEM_ERROR("CcoWindowRegister: P2P FD exchange failed");
+      localBoot.Finalize();
+      return -1;
+    }
+
+    std::vector<int> globalToPeer(worldSize, -1);
+    for (int i = 0; i < p2pWorldSize; i++) {
+      globalToPeer[sortedGroup[i]] = i;
+    }
+
+    hipMemAccessDesc accessDesc = {};
+    accessDesc.location.type = hipMemLocationTypeDevice;
+    accessDesc.location.id = currentDev;
+    accessDesc.flags = hipMemAccessFlagsProtReadWrite;
+
+    for (int pe : p2pPeers) {
+      int pr = globalToPeer[pe];
+      if (pr < 0 || pr >= static_cast<int>(allFds.size())) continue;
+      int peerFd = allFds[pr][0];
+      if (peerFd < 0) continue;
+
+      hipMemGenericAllocationHandle_t importedHandle;
+      hipError_t err = hipMemImportFromShareableHandleCompat(
+          &importedHandle, peerFd, hipMemHandleTypePosixFileDescriptor);
+      if (err != hipSuccess) {
+        MORI_SHMEM_WARN("CcoWindowRegister: import from PE {} failed: {}", pe, err);
+        continue;
+      }
+
+      int peerLsaRank = pe - comm->myNodeStart;
+      void* peerVa = static_cast<char*>(comm->flatBase) +
+                     static_cast<size_t>(peerLsaRank) * comm->perRankSize + slotOffset;
+      HIP_RUNTIME_CHECK(hipMemMap(peerVa, alignedSize, 0, importedHandle, 0));
+
+      // hipMemSetAccess can transiently fail under concurrent VMM operations.
+      for (int retry = 0;; retry++) {
+        hipError_t setErr = hipMemSetAccess(peerVa, alignedSize, &accessDesc, 1);
+        if (setErr == hipSuccess) break;
+        if (retry >= 5) { HIP_RUNTIME_CHECK(setErr); }
+        usleep(1000 * (1 << retry));
+      }
+    }
+
+    localBoot.Finalize();
+  }
+
+  // RDMA MR registration + rkey Allgather.
+  uint32_t lkey = 0;
+  uint32_t localRkey = 0;
+
+  application::RdmaDeviceContext* rdmaDevCtx = comm->ctx->GetRdmaDeviceContext();
+  if (rdmaDevCtx && shareFd >= 0) {
+    application::RdmaMemoryRegion mr;
+    if (comm->iovaZeroMode) {
+      mr = rdmaDevCtx->RegisterRdmaMemoryRegionDmabufIova0(localPtr, size, shareFd);
+    } else {
+      mr = rdmaDevCtx->RegisterRdmaMemoryRegionDmabuf(localPtr, size, shareFd);
+    }
+    lkey = mr.lkey;
+    localRkey = mr.rkey;
+  }
+
+  auto* peerRkeys_host = static_cast<uint32_t*>(calloc(worldSize, sizeof(uint32_t)));
+  peerRkeys_host[rank] = localRkey;
+  comm->bootNet->Allgather(&localRkey, peerRkeys_host, sizeof(uint32_t));
+
+  // SDMA signals — intra-node only, sized lsaSize × sdmaNumQueue. The IPC
+  // handle Allgather still spans worldSize (avoids a second bootstrap); we
+  // only open the LSA peers' handles into the lsa-sized peerSignalPtrs.
+  int sdmaNumQueue = comm->sdmaNumQueue;
+  size_t signalArraySize = static_cast<size_t>(comm->lsaSize) * sdmaNumQueue * sizeof(HSAuint64);
+
+  HSAuint64* signalPtrs = nullptr;
+  HSAuint64* expectSignalsPtr = nullptr;
+  HSAuint64** peerSignalPtrs_host_arr = nullptr;
+  HSAuint64** peerSignalPtrs_gpu = nullptr;
+
+  // Materialize SDMA signals only if the user opted in (policy) AND at least
+  // one peer is SDMA-capable (hardware).
+  bool hasSdmaPeers = false;
+  if (comm->ctx->IsSdmaEnabled()) {
+    for (int pe = 0; pe < worldSize; pe++) {
+      if (comm->ctx->GetPeerCapabilities(pe).canSDMA) {
+        hasSdmaPeers = true;
+        break;
+      }
+    }
+  }
+
+  if (hasSdmaPeers && sdmaNumQueue > 0) {
+    HIP_RUNTIME_CHECK(hipMalloc(&signalPtrs, signalArraySize));
+    HIP_RUNTIME_CHECK(hipMemset(signalPtrs, 0, signalArraySize));
+    HIP_RUNTIME_CHECK(hipMalloc(&expectSignalsPtr, signalArraySize));
+    HIP_RUNTIME_CHECK(hipMemset(expectSignalsPtr, 0, signalArraySize));
+
+    hipIpcMemHandle_t signalHandle;
+    HIP_RUNTIME_CHECK(hipIpcGetMemHandle(&signalHandle, signalPtrs));
+
+    auto* signalHandles =
+        static_cast<hipIpcMemHandle_t*>(calloc(worldSize, sizeof(hipIpcMemHandle_t)));
+    comm->bootNet->Allgather(&signalHandle, signalHandles, sizeof(hipIpcMemHandle_t));
+
+    peerSignalPtrs_host_arr =
+        static_cast<HSAuint64**>(calloc(comm->lsaSize, sizeof(HSAuint64*)));
+    peerSignalPtrs_host_arr[comm->lsaRank] = signalPtrs;
+    for (int lsa = 0; lsa < comm->lsaSize; lsa++) {
+      if (lsa == comm->lsaRank) continue;
+      int pe = comm->myNodeStart + lsa;
+      if (!comm->ctx->GetPeerCapabilities(pe).canSDMA) continue;
+      void* mapped = nullptr;
+      HIP_RUNTIME_CHECK(
+          hipIpcOpenMemHandle(&mapped, signalHandles[pe], hipIpcMemLazyEnablePeerAccess));
+      peerSignalPtrs_host_arr[lsa] = reinterpret_cast<HSAuint64*>(mapped);
+    }
+
+    HIP_RUNTIME_CHECK(hipMalloc(&peerSignalPtrs_gpu, sizeof(HSAuint64*) * comm->lsaSize));
+    HIP_RUNTIME_CHECK(hipMemcpy(peerSignalPtrs_gpu, peerSignalPtrs_host_arr,
+                                sizeof(HSAuint64*) * comm->lsaSize, hipMemcpyHostToDevice));
+    free(signalHandles);
+  }
+
+  uint32_t* peerRkeys_gpu = nullptr;
+  HIP_RUNTIME_CHECK(hipMalloc(&peerRkeys_gpu, sizeof(uint32_t) * worldSize));
+  HIP_RUNTIME_CHECK(hipMemcpy(peerRkeys_gpu, peerRkeys_host, sizeof(uint32_t) * worldSize,
+                              hipMemcpyHostToDevice));
+
+  CcoWindowDevice hostShadow = {};
+  hostShadow.winBase = static_cast<char*>(comm->flatBase) + slotOffset;
+  hostShadow.stride4G = static_cast<uint32_t>(comm->perRankSize >> 32);
+  hostShadow.lsaRank = comm->lsaRank;
+  hostShadow.ibgdaWin.peerRkeys = peerRkeys_gpu;
+  hostShadow.ibgdaWin.lkey = lkey;
+  hostShadow.deviceHandles_d = comm->sdmaDevHandles;
+  hostShadow.signalPtrs = signalPtrs;
+  hostShadow.expectSignalsPtr = expectSignalsPtr;
+  hostShadow.peerSignalPtrs = peerSignalPtrs_gpu;
+  hostShadow.sdmaNumQueue = static_cast<uint32_t>(sdmaNumQueue);
+
+  CcoWindowDevice* devPtr = nullptr;
+  HIP_RUNTIME_CHECK(hipMalloc(&devPtr, sizeof(CcoWindowDevice)));
+  HIP_RUNTIME_CHECK(
+      hipMemcpy(devPtr, &hostShadow, sizeof(CcoWindowDevice), hipMemcpyHostToDevice));
+
+  // Publish into the per-comm window table (drives findWindow lookups).
+  CcoComm::WindowTableEntry tableEntry;
+  tableEntry.base = reinterpret_cast<uintptr_t>(localPtr);
+  tableEntry.size = static_cast<uintptr_t>(size);
+  tableEntry.devPtr = devPtr;
+  comm->windowTableEntries.push_back(tableEntry);
+
+  auto* wh = new CcoWindowHost();
+  wh->localPtr = localPtr;
+  wh->size = size;
+  wh->signalPtrs = signalPtrs;
+  wh->expectSignalsPtr = expectSignalsPtr;
+  wh->peerSignalPtrs = peerSignalPtrs_host_arr;
+  wh->devPtr = devPtr;
+  wh->peerRkeys_gpu = peerRkeys_gpu;
+  wh->peerSignalPtrs_gpu = peerSignalPtrs_gpu;
+  comm->windows.push_back(wh);
+
+  *outWin = devPtr;
+
+  char* winBase = static_cast<char*>(comm->flatBase) + slotOffset;
+  MORI_SHMEM_INFO("CcoWindowRegister: rank={} win={} winBase={} size={} slotOffset={} lkey={}",
+                  rank, (void*)devPtr, (void*)winBase, size, slotOffset, lkey);
+  for (int lsa = 0; lsa < comm->lsaSize; lsa++) {
+    int pe = comm->myNodeStart + lsa;
+    void* peerVa = winBase + static_cast<size_t>(lsa) * comm->perRankSize;
+    MORI_SHMEM_INFO("  LSA[{}] (PE {}): flatVA={} rkey={}", lsa, pe, peerVa, peerRkeys_host[pe]);
+  }
+  for (int pe = 0; pe < worldSize; pe++) {
+    if (pe >= comm->myNodeStart && pe < comm->myNodeStart + comm->lsaSize) continue;
+    MORI_SHMEM_INFO("  XNODE PE {}: rkey={} (RDMA via iova=0)", pe, peerRkeys_host[pe]);
+  }
+  if (signalPtrs) {
+    MORI_SHMEM_INFO("  SDMA: signalPtrs={} expectSignals={} numQueue={}",
+                    (void*)signalPtrs, (void*)expectSignalsPtr, sdmaNumQueue);
+  }
+  MORI_SHMEM_INFO("  deviceHandles_d={}", (void*)comm->sdmaDevHandles);
+
+  free(peerRkeys_host);
+
+  return 0;
+}
+
+/* ========================================================================== */
+/*                      CcoWindowRegister (convenience)                    */
+/* ========================================================================== */
+
+int CcoWindowRegister(CcoComm* comm, size_t size, CcoWindow_t* outWin, void** localPtr) {
+  void* ptr = nullptr;
+  int ret = CcoMemAlloc(comm, size, &ptr);
+  if (ret != 0) return ret;
+
+  ret = CcoWindowRegister(comm, ptr, size, outWin);
+  if (ret != 0) {
+    CcoMemFree(comm, ptr);
+    return ret;
+  }
+
+  *localPtr = ptr;
+  return 0;
+}
+
+/* ========================================================================== */
+/*                          CcoWindowDeregister                            */
+/* ========================================================================== */
+
+int CcoWindowDeregister(CcoComm* comm, CcoWindow_t win) {
+  CcoWindowHost* wh = nullptr;
+  size_t idx = 0;
+  for (size_t i = 0; i < comm->windows.size(); i++) {
+    if (comm->windows[i]->devPtr == win) {
+      wh = comm->windows[i];
+      idx = i;
+      break;
+    }
+  }
+  if (!wh) {
+    MORI_SHMEM_WARN("CcoWindowDeregister: win {} not found", (void*)win);
+    return -1;
+  }
+
+  MORI_SHMEM_TRACE("CcoWindowDeregister: rank={} ptr={}", comm->rank, wh->localPtr);
+
+  // Unmap the P2P peer slots that WindowRegister mapped (ENOMAP is fine).
+  auto allocIt = comm->allocTable.find(wh->localPtr);
+  if (allocIt != comm->allocTable.end()) {
+    size_t slotOff = allocIt->second.slotOffset;
+    size_t allocSize = allocIt->second.size;
+    for (int lsa = 0; lsa < comm->lsaSize; lsa++) {
+      if (lsa == comm->lsaRank) continue;
+      int pe = comm->myNodeStart + lsa;
+      if (!comm->ctx->CanUseP2P(pe)) continue;
+      void* peerVa = static_cast<char*>(comm->flatBase) +
+                     static_cast<size_t>(lsa) * comm->perRankSize + slotOff;
+      (void)hipMemUnmap(peerVa, allocSize);
+    }
+  }
+
+  auto& entries = comm->windowTableEntries;
+  entries.erase(std::remove_if(entries.begin(), entries.end(),
+                               [win](const CcoComm::WindowTableEntry& e) {
+                                 return e.devPtr == win;
+                               }),
+                entries.end());
+
+  application::RdmaDeviceContext* rdmaDevCtx = comm->ctx->GetRdmaDeviceContext();
+  if (rdmaDevCtx) rdmaDevCtx->DeregisterRdmaMemoryRegion(wh->localPtr);
+
+  if (wh->peerRkeys_gpu) HIP_RUNTIME_CHECK(hipFree(wh->peerRkeys_gpu));
+  if (wh->signalPtrs) HIP_RUNTIME_CHECK(hipFree(wh->signalPtrs));
+  if (wh->expectSignalsPtr) HIP_RUNTIME_CHECK(hipFree(wh->expectSignalsPtr));
+  if (wh->peerSignalPtrs_gpu) HIP_RUNTIME_CHECK(hipFree(wh->peerSignalPtrs_gpu));
+  if (wh->devPtr) HIP_RUNTIME_CHECK(hipFree(wh->devPtr));
+
+  free(wh->peerSignalPtrs);
+
+  comm->windows.erase(comm->windows.begin() + idx);
+  delete wh;
   return 0;
 }
